Finished CLI and configCLI
